moved plugin interface into kernel package
