+++ pkg/kernel/plugin.go
@@ -1,10 +1,10 @@
-package interfaces
+package kernel
 
 type Plugin interface {
 	ID() string
 	Dependencies() []string
 	Capabilities() []string
-	Initialize() error
+	Initialize(api API) error
 	Start() error
 	Stop() error
 }
